LayoutTests/fast/js/codegen-temporaries-expected.txt

Tests whether bytecode codegen properly handles temporaries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a is true
PASS b is false
PASS successfullyParsed is true

TEST COMPLETE
